[tests] Split fooService/FooServiceJTA tests
